Flower/requirements.txt
asgiref==3.8.1
Django==4.2
Django==4.2
djangorestframework==3.14.0
django-filter==24.2
djangorestframework-simplejwt==5.2.2
django-cors-headers==4.3.1
PyJWT==2.8.0
pytz==2024.1
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1